[fe] refact: NoticeBoardList에서 추가적으로 Component화 진행
